Moved project fetching to utils
